Isolate 2 tests that need porting to Py3 later #295
