+++ tests/scancode/test_cli.py
@@ -634,22 +634,6 @@ def test_scan_progress_display_is_not_damaged_with_long_file_names(monkeypatch):
         assert expected3 not in result.output
 
 
-def test_scan_does_scan_php_composer():
-    test_file = test_env.get_test_loc('composer/composer.json')
-    expected_file = test_env.get_test_loc('composer/composer.expected.json')
-    result_file = test_env.get_temp_file('results.json')
-    run_scan_click(['--package', test_file, '--json', result_file])
-    check_json_scan(expected_file, result_file)
-
-
-def test_scan_does_scan_rpm():
-    test_file = test_env.get_test_loc('rpm/fping-2.4-0.b2.rhfc1.dag.i386.rpm')
-    expected_file = test_env.get_test_loc('rpm/fping-2.4-0.b2.rhfc1.dag.i386.rpm.expected.json')
-    result_file = test_env.get_temp_file('results.json')
-    run_scan_click(['--package', test_file, '--json', result_file])
-    check_json_scan(expected_file, result_file, regen=False)
-
-
 def test_scan_cli_help(regen=False):
     expected_file = test_env.get_test_loc('help/help.txt')
     result = run_scan_click(['--help'])
