+++ tests/scancode/test_cli2.py
@@ -1,0 +1,60 @@
+#
+# Copyright (c) 2018 nexB Inc. and others. All rights reserved.
+# http://nexb.com and https://github.com/nexB/scancode-toolkit/
+# The ScanCode software is licensed under the Apache License version 2.0.
+# Data generated with ScanCode require an acknowledgment.
+# ScanCode is a trademark of nexB Inc.
+#
+# You may not use this software except in compliance with the License.
+# You may obtain a copy of the License at: http://apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+#
+# When you publish or redistribute any data created with ScanCode or any ScanCode
+# derivative work, you must accompany this data with the following acknowledgment:
+#
+#  Generated with ScanCode and provided on an "AS IS" BASIS, WITHOUT WARRANTIES
+#  OR CONDITIONS OF ANY KIND, either express or implied. No content created from
+#  ScanCode should be considered or used as legal advice. Consult an Attorney
+#  for any legal advice.
+#  ScanCode is a free software code scanning tool from nexB Inc. and others.
+#  Visit https://github.com/nexB/scancode-toolkit/ for support and download.
+
+from __future__ import absolute_import
+from __future__ import print_function
+from __future__ import division
+from __future__ import unicode_literals
+
+import os
+
+import click
+click.disable_unicode_literals_warning = True
+
+from commoncode.testcase import FileDrivenTesting
+
+from scancode.cli_test_utils import check_json_scan
+from scancode.cli_test_utils import run_scan_click
+
+
+# TODO: port to Python 3
+
+test_env = FileDrivenTesting()
+test_env.test_data_dir = os.path.join(os.path.dirname(__file__), 'data')
+
+
+def test_scan_does_scan_php_composer():
+    test_file = test_env.get_test_loc('composer/composer.json')
+    expected_file = test_env.get_test_loc('composer/composer.expected.json')
+    result_file = test_env.get_temp_file('results.json')
+    run_scan_click(['--package', test_file, '--json', result_file])
+    check_json_scan(expected_file, result_file)
+
+
+def test_scan_does_scan_rpm():
+    test_file = test_env.get_test_loc('rpm/fping-2.4-0.b2.rhfc1.dag.i386.rpm')
+    expected_file = test_env.get_test_loc('rpm/fping-2.4-0.b2.rhfc1.dag.i386.rpm.expected.json')
+    result_file = test_env.get_temp_file('results.json')
+    run_scan_click(['--package', test_file, '--json', result_file])
+    check_json_scan(expected_file, result_file, regen=False)
